Refactor user follow/unfollow functionality; implement UserFollow entity and update UserService methods

+++ src/main/java/blog/code/codeblog/dto/user/UserResponseDTO.java
@@ -10,6 +10,6 @@ public record UserResponseDTO(
         String name,
         String login,
         String urlProfilePic,
-        int followersCount,
-        int followingCount
+        long followersCount,
+        long followingCount
 ) {}
